test: add full ERC20 blackbox
